backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
email-validator>=2.0.0
motor>=3.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
networkx>=3.0.0
certifi>=2023.7.22
google-generativeai>=0.3.0
